feat: configure anvil to listen on non-local address
For use in a remote network, Anvil must listen on the public IP address of the VM it is running on.
The `anvil` binary respects the `ANVIL_IP_ADDR` variable for configuring a listener address.
However, it turns out the `alloy` crate is actually hard coded to work with "localhost", so
unfortunately the `AnvilInstance::endpoint` function cannot be used to return the resulting RPC URL.
For that reason, a new `rpc_url` field was added to the `Testnet` struct to keep track of the
non-local listening address.

The deployer wallet private key was also saved in the CSV data because it will be necessary for
`testnet-deploy` to obtain this.