feat: select evm network based on network id
The doc comments on the `get_evm_network` function provide more detail, but the summary is:
* The `EVM_NETWORK` and associated variables are checked first. They allow the selection of a preset
  or a custom configuration.
* Otherwise if these variables are not being used:
    + If no network ID is being used and it's not a local network, we use network ID `1` and so we
      select Arbitrum One.
    + If the `--alpha` argument has been used we use network ID `2` and so we select Arbitrum
      Sepolia Test.
    + If a network ID has been explicitly supplied we are in a testnet scenario and therefore we
      will assert that the `EVM_NETWORK` variables must be used.

If these conditions are not met the binary will error. I think this is better than selecting a
default network because when I've been testing with users I've seen this result in confusion about
exactly which EVM network was being selected.