fix: correctly obtain evm network for `wallet balance`
When the `--alpha` argument was introduced, this command was not updated to provide the network ID,
which is a factor when obtaining the EVM provider.

This meant when the `EVM_NETWORK` argument was not used, the command would error.

It now works as expected, where not explicitly setting the EVM network will result in Arbitrum One
being used.